- Removed extraneous remnants - Added ability to declare multiple tuners. Still only uses the first tuner specified though as I only have one available to test and db considerations need to be had.
to specify set 'tuners' paramater in [cfg] section of ini file to have a space separated list of ini sections. They can be whatever, it does not matter. Then for each section identified, a section *must* be declared in the ini file with the tunerid parameter and a tuners parameter that declares a space separated list of tuners available for that tunerid.

eg:

[cfg]
tuners=tuner1 othertuner

[tuner1]
tunerid=123n4411
tuners=/tuner0/

[othertuner]
tunerid=1293n321
tuners =/tuner0/ /tuner1/