Created a Ruby on Rails branch of the web application.
Rails makes it easy to manage changes to the interface by using the MVC
pattern. It also automates a lot of things, including scaffolds which
make it easy to create administrative interfaces (sometimes)
automatically. This make it much easier to create all the boring
administrative views that I never seemed inclined to create before.

Rails uses the ActiveRecord gem to enable database independence, so the
app is no longer tied to MySQL, like before. Also, this will
allow the web app and the script to share database models, which means
that changes to the database are largely shielded from the perl script.
Of course, to share the model with the rails app, the perl script must
be re-written in ruby.

I have largely re-written the perl script as script/scan. To run it, cd
to script and do ./runner scan. This will load the rails runtime and all
the models and run the script.

I have created a Station model, but I have not yet created models for
logging results or for tuners, so right now the script just inserts any
stations that it finds. I have also added lib/haversine.rb so to
calculate great circle distances, since ruby doesn't have a built-in
great circle distance function like perl. It is largely unchanged from
http://www.esawdust.com/blog/gps/files/HaversineFormulaInRuby.html.