Added Tuner and TunerInfo models.
A Tuner consists of a tuner_id and tuner_number (eg. 1016D42E, 0).
TunerInfo contains additional information about what that tuner is
connected to. Currently this only has one field: a free-form text field.

Tuner fields are not allowed to be updated by the web UI. Tuner
information can be 'updated,' though. Whenever a Tuner is 'updated,' a
new TunerInfo is created and added to the list of TunerInfos on Tuner.
This keeps a log of when a Tuner was 'updated.' This will allow
corresponding log entries to TunerInfo, so that changes in tuner
configurations can be correlated to log entries.

Example:
A user gets a new antenna, which has better receive gain. After
switching out the antenna, he goes in to the web UI and updates the
tuner info to now show that he is using this new antenna. His log results
for this tuner show dramatic improvement. Viewing the timeline of log
results will show the tuner info update event and a correlation can be
made between when he changed out his antenna and when his results
improved.

Example:
A user decides to rotate her antenna to a different direction. After
rotating the antenna, she updates the web UI to note that the antenna is
now pointing South. The scanner detects new stations and logs them.
Looking at the timeline, she can see that when she rotated the antenna
she started seeing these new results.

Other fields could be added to TunerInfo, such as direction, antenna,
height, etc.