Added signal graphing functions using Highcharts (www.highcharts.com).
Note: This is licenced under the Creative Commons Attribution -
NonCommercial 3.0 licence only for non commercial use.

The package is very nice and has a lot of functionality and is very
shiny, but it does not do exactly what I want. All the points are
assumed continuous (which my data is not) and evenly spaced (also not).
To work around this, I have to fill in the gaps and it shows on the
graph with lines that go to 0. What I actually want is for the lines to
become disconnected instead of going to 0, which more accurately
reflects the data.

I plan on using the Raphael javascript svg drawing library for this,
which is MIT licenced, which is much nicer. Unfortunately, it is just a
primitive drawing library. There is an example of a simple chart, so I
will use this as the basis for my javascript chart library/functions.

For now, Highcharts will work fine.