Create caching table for recent log entries
Instead of querying through all the logs to find the latest log for
each station and tuner, which is only really easy to do with postgres,
and slow in any case, we keep a cache of the latest log entry in the
RecentLog model. By adding an after_save callback on the Log model,
which creates or updates a RecentLog entry, this will always be
consistent.