Merge TunerInfo in to Tuner
Instead of using a half-baked homebrew versioning system, migrate
to an existing solution. This is step 1. Since there are only a
couple deployed applications with hardly any history worth saving,
it was deemed not worth the effort to enable vestal_versions prior
to dropping the old information. Step 2 will be to enable
vestal_versions (or some other versioning gem).